Refactoring of Phase1 OMTF emulator code

+++ L1Trigger/L1TMuonOverlapPhase1/interface/Omtf/GoldenPatternWithStat.h
@@ -14,7 +14,7 @@ class OMTFConfiguration;
 class GoldenPatternWithStat : public GoldenPatternWithThresh {
 public:
   static const unsigned int STAT_BINS = 1;  //TODO change value if needed
-  typedef boost::multi_array<float, 4> StatArrayType;
+  typedef boost::multi_array<int, 4> StatArrayType;
 
   GoldenPatternWithStat(const Key& aKey, unsigned int nLayers, unsigned int nRefLayers, unsigned int nPdfAddrBits);
